Increase max_http_header_data to fix proxy issues
Libwebsockets was issuing an error "Ran out of header data space"
when a server was running behind an authentication proxy.

The size of the "Authorization: Negotiate <token>" header can be
very long, potentially depending on the number of groups a user is
a member of (https://support.microsoft.com/en-us/help/2020943).

Doubling the max_http_header_data from 4096 (default) to 8192
fixes the problem (in the BBP setup).